Fix tests

+++ src/tokens/FxGenArt721.sol
@@ -282,7 +282,7 @@ contract FxGenArt721 is IFxGenArt721, Initializable, ERC721, Ownable, Pausable, 
      * @dev Registers arbitrary number of minter contracts
      * @param _mintInfo List of minter contracts and their reserves
      */
-    function _registerMinters(address _owner, uint128 _lockTime, MintInfo[] calldata _mintInfo) internal {
+    function _registerMinters(address _owner, uint256 _lockTime, MintInfo[] calldata _mintInfo) internal {
         address minter;
         uint128 totalAllocation;
         ReserveInfo memory reserveInfo;
